# packages/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/timetable/groups": {
      "get": {
        "description": "Comes from the synchronised catalogue, never from a live upstream call. Identifiers are Campus UUIDs; the source system id is never exposed.",
        "operationId": "TimetableController_groups_v1",
        "parameters": [
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "de",
                "en"
              ],
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Max 50.",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "department",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "description": "Searches short name, long name and department.",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimetableGroupsResponseDto"
                }
              }
            }
          }
        },
        "summary": "List selectable class groups.",
        "tags": [
          "timetable"
        ]
      }
    },
    "/v1/timetable/groups/{groupId}": {
      "get": {
        "description": "Returns one group by its stable Campus UUID so clients can retain the selection and its public label offline without downloading the catalogue.",
        "operationId": "TimetableController_group_v1",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "description": "Stable Campus UUID.",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimetableGroupResponseDto"
                }
              }
            }
          }
        },
        "summary": "Resolve one previously selected class group.",
        "tags": [
          "timetable"
        ]
      }
    },
    "/v1/timetable/periods": {
      "get": {
        "description": "Contains Campus UUIDs only. WebUntis school-year and class identifiers remain server-internal.",
        "operationId": "TimetableController_periods_v1",
        "parameters": [
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "de",
                "en"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimetablePeriodsResponseDto"
                }
              }
            }
          }
        },
        "summary": "List semester catalogues available to the timetable assistant.",
        "tags": [
          "timetable"
        ]
      }
    },
    "/v1/timetable/lesson-info": {
      "get": {
        "description": "Reads the last successfully imported timetable window from the Campus database. No WebUntis request is triggered.",
        "operationId": "TimetableController_lessonInfo_v1",
        "parameters": [
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "de",
                "en"
              ],
              "type": "string"
            }
          },
          {
            "name": "groupId",
            "required": true,
            "in": "query",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimetableLessonInfoResponseDto"
                }
              }
            }
          }
        },
        "summary": "List exact lesson information texts for a selected timetable group.",
        "tags": [
          "timetable"
        ]
      }
    },
    "/v1/timetable/modules": {
      "get": {
        "description": "Reads retained Campus timetable rows only. Module names and codes are source-provided; no upstream identifiers or live request are exposed.",
        "operationId": "TimetableController_modules_v1",
        "parameters": [
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "de",
                "en"
              ],
              "type": "string"
            }
          },
          {
            "name": "groupId",
            "required": true,
            "in": "query",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimetableModulesResponseDto"
                }
              }
            }
          }
        },
        "summary": "List modules observed for one selectable study group.",
        "tags": [
          "timetable"
        ]
      }
    },
    "/v1/timetable/entries": {
      "get": {
        "description": "Every day of the range is returned, so a genuinely free day is distinguishable from a loading error. Range is capped at 42 days.",
        "operationId": "TimetableController_entries_v1",
        "parameters": [
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "de",
                "en"
              ],
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "example": "2026-08-02"
            }
          },
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "example": "2026-07-20"
            }
          },
          {
            "name": "groupId",
            "required": true,
            "in": "query",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimetableWeekResponseDto"
                }
              }
            }
          }
        },
        "summary": "Fetch one group’s timetable for a date range.",
        "tags": [
          "timetable"
        ]
      }
    },
    "/v1/timetable/status": {
      "get": {
        "description": "Deliberately thin: no upstream URLs, headers, external ids or error detail.",
        "operationId": "TimetableController_status_v1",
        "parameters": [
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "de",
                "en"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimetableStatusResponseDto"
                }
              }
            }
          }
        },
        "summary": "Public health of the timetable integration.",
        "tags": [
          "timetable"
        ]
      }
    },
...
  },
  "info": {
    "title": "Campus Köthen API",
    "description": "Public read-only API of the Campus Köthen app.\n\nCampus Köthen is an independent, unofficial campus app. It is neither developed\nnor operated by Hochschule Anhalt, nor is it officially endorsed by the university.\n\nAll content endpoints accept `locale=de|en` (default `de`) and report\n`requestedLocale`, `resolvedLocale` and `translationFallback` in their metadata.\nCanteen dish text comes from a German-only source and is never machine-translated.",
    "version": "1.2.4",
    "contact": {
      "name": "Campus Köthen App",
      "url": "https://github.com/Leviora-Studio/campus-koethen",
      "email": ""
    },
    "license": {
      "name": "AGPL-3.0-only",
      "url": "https://www.gnu.org/licenses/agpl-3.0.html"
    }
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 20
          },
          "total": {
            "type": "number",
            "example": 42
          },
          "totalPages": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "totalPages"
        ]
      },
      "ResponseMetaDto": {
        "type": "object",
        "properties": {
          "requestedLocale": {
            "type": "string",
            "enum": [
              "de",
              "en"
            ],
            "description": "Locale the client asked for."
          },
          "resolvedLocale": {
            "type": "string",
            "enum": [
              "de",
              "en"
            ],
            "description": "Locale actually served."
          },
          "translationFallback": {
            "type": "boolean",
            "description": "True when at least one delivered field came from the German fallback although another locale was requested."
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          },
          "droppedBlockTypes": {
            "description": "Content block types removed because the MVP client does not render them.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "lastSuccessfulSyncAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Last successful canteen synchronisation. Null means: never succeeded."
          },
          "dataStale": {
            "type": "boolean",
            "description": "True when the data is older than the configured threshold."
          },
          "from": {
            "type": "string",
            "description": "Start of the resolved date range (YYYY-MM-DD)."
          },
          "to": {
            "type": "string",
            "description": "End of the resolved date range (YYYY-MM-DD)."
          },
          "maxRangeDays": {
            "type": "number",
            "description": "Maximum inclusive number of days in one calendar events request."
          },
          "truncated": {
            "type": "boolean",
            "description": "True when the server hit its own result ceiling and deliberately cut the list short. The client must say so rather than present the result as complete."
          },
          "featureEnabled": {
            "type": "boolean",
            "description": "False when the timetable integration is switched off server-side."
          },
          "dataState": {
            "type": "string",
            "enum": [
              "ready",
              "pending",
              "unavailable"
            ],
            "description": "ready = data present; pending = enabled but not yet synchronised; unavailable = disabled or permanently without data."
          },
          "timezone": {
            "type": "string",
            "description": "Business timezone the wall-clock times belong to.",
            "example": "Europe/Berlin"
          }
        },
        "required": [
          "requestedLocale",
          "resolvedLocale",
          "translationFallback"
        ]
      },
...
      "TimetableGroupDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Stable Campus identifier."
          },
          "shortName": {
            "type": "string",
            "example": "AIN2 - BT"
          },
          "longName": {
            "type": "string",
            "example": "AIN2-Angewandte Informatik Vertiefung: Biotechnologie"
          },
          "department": {
            "type": "string",
            "nullable": true,
            "example": "FB5"
          }
        },
        "required": [
          "id",
          "shortName",
          "longName",
          "department"
        ]
      },
      "TimetableGroupsResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimetableGroupDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "TimetableGroupResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TimetableGroupDto"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "TimetablePeriodDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Stable Campus identifier."
          },
          "name": {
            "type": "string",
            "description": "Source-provided semester label. Never translated."
          },
          "validFrom": {
            "type": "string",
            "format": "date",
            "example": "2026-10-05"
          },
          "validTo": {
            "type": "string",
            "format": "date",
            "example": "2027-03-31"
          },
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimetableGroupDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "validFrom",
          "validTo",
          "groups"
        ]
      },
      "TimetablePeriodsResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimetablePeriodDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "TimetableLessonInfoDto": {
        "type": "object",
        "properties": {
          "values": {
            "description": "Distinct lessonInfo texts exactly as stored for the selected group.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "hasWithoutInfo": {
            "type": "boolean",
            "description": "Whether the covered timetable also contains entries without lessonInfo."
          }
        },
        "required": [
          "values",
          "hasWithoutInfo"
        ]
      },
      "TimetableLessonInfoResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TimetableLessonInfoDto"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "TimetableModuleDto": {
        "type": "object",
        "properties": {
          "subjectCode": {
            "type": "string",
            "nullable": true,
            "description": "Source-provided module code."
          },
          "title": {
            "type": "string",
            "description": "Source-provided module title. Never translated."
          }
        },
        "required": [
          "subjectCode",
          "title"
        ]
      },
      "TimetableModulesResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimetableModuleDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "TimetableTeacherDto": {
        "type": "object",
        "properties": {
          "shortName": {
            "type": "string",
            "description": "Name as published by the source. Never translated."
          },
          "displayName": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "shortName",
          "displayName"
        ]
      },
      "TimetableRoomDto": {
        "type": "object",
        "properties": {
          "shortName": {
            "type": "string"
          },
          "longName": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "shortName",
          "longName"
        ]
      },
      "TimetableEntryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "start": {
            "type": "string",
            "format": "date-time",
            "description": "Absolute instant, UTC."
          },
          "end": {
            "type": "string",
            "format": "date-time"
          },
          "timezone": {
            "type": "string",
            "example": "Europe/Berlin",
            "description": "Business timezone the source schedules in."
          },
          "title": {
            "type": "string",
            "description": "Subject name from the source. Never translated."
          },
          "subjectCode": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "enum": [
              "regular_teaching",
              "additional",
              "unknown"
            ],
            "description": "Normalised. An unrecognised upstream value becomes `unknown`."
          },
          "status": {
            "type": "string",
            "enum": [
              "regular",
              "changed",
              "cancelled",
              "unknown"
            ],
            "description": "Normalised. The client must render `unknown` neutrally rather than guess."
          },
          "teachers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimetableTeacherDto"
            }
          },
          "rooms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimetableRoomDto"
            }
          },
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimetableGroupDto"
            }
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "lessonInfo": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "start",
          "end",
          "timezone",
          "title",
          "subjectCode",
          "type",
          "status",
          "teachers",
          "rooms",
          "groups",
          "note",
          "lessonInfo"
        ]
      },
      "TimetableDayDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "example": "2026-07-20"
          },
          "entries": {
            "description": "Empty means a genuinely free day, not a loading failure.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimetableEntryDto"
            }
          }
        },
        "required": [
          "date",
          "entries"
        ]
      },
      "TimetableWeekDto": {
        "type": "object",
        "properties": {
          "group": {
            "$ref": "#/components/schemas/TimetableGroupDto"
          },
          "days": {
            "description": "Every day of the requested range.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimetableDayDto"
            }
          }
        },
        "required": [
          "group",
          "days"
        ]
      },
      "TimetableWeekResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TimetableWeekDto"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "TimetableStatusDto": {
        "type": "object",
        "properties": {
          "featureEnabled": {
            "type": "boolean"
          },
          "groupCount": {
            "type": "number",
            "example": 270
          },
          "lastGroupSyncAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "lastEntrySyncAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "dataStale": {
            "type": "boolean"
          },
          "coveredFrom": {
            "type": "string",
            "nullable": true,
            "example": "2026-07-15"
          },
          "coveredTo": {
            "type": "string",
            "nullable": true,
            "example": "2026-08-19"
          }
        },
        "required": [
          "featureEnabled",
          "groupCount",
          "lastGroupSyncAt",
          "lastEntrySyncAt",
          "dataStale",
          "coveredFrom",
          "coveredTo"
        ]
      },
      "TimetableStatusResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TimetableStatusDto"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
    }
  }
}